+++ apps/router-e2e/__e2e__/01-rsc/app/_layout.tsx
@@ -1,12 +1,25 @@
 import 'server-only';
-import { Children } from 'expo-router/build/rsc/router/host';
 
 import { View } from '../lib/react-native';
 
-const HomeLayout = () => {
+import '../global.css';
+import { unstable_styles } from '../home.module.css';
+
+const HomeLayout = (props) => {
   return (
     <View style={{ flex: 1, padding: 12 }} testID="layout-child-wrapper">
-      <Children />
+      {props.children}
+      <View
+        testID="layout-global-style"
+        style={[
+          { width: 100, height: 100 },
+          { $$css: true, _: 'custom-global-style' },
+        ]}
+      />
+      <View
+        testID="layout-module-style"
+        style={[{ width: 100, height: 100 }, unstable_styles.container]}
+      />
     </View>
   );
 };