+++ packages/@expo/metro-config/src/__tests__/babel-transformer.test.ts
@@ -66,6 +66,7 @@ it(`passes the environment as isServer to the babel preset`, () => {
       isDev: true,
       bundler: 'metro',
       engine: undefined,
+      isDOM: false,
       name: 'metro',
       platform: 'ios',
       baseUrl: '',