+++ packages/expo-camera/ios/Current/CameraView.swift
@@ -398,6 +398,10 @@ public class CameraView: ExpoView, EXCameraInterface, EXAppLifecycleListener,
   }
 
   public func photoOutput(_ output: AVCapturePhotoOutput, willCapturePhotoFor resolvedSettings: AVCaptureResolvedPhotoSettings) {
+    if photoCaptureOptions?.shutterSound == false {
+      AudioServicesDisposeSystemSoundID(1108)
+    }
+
     guard animateShutter else {
       return
     }