+++ packages/expo-font/ios/FontLoaderModule.swift
@@ -33,16 +33,17 @@ public final class FontLoaderModule: Module {
       }
 
       // Register the font
-      try registerFont(fontUrl)
+      try registerFont(fontUrl: fontUrl, fontFamilyAlias: fontFamilyAlias)
 
       // Create a font object from the given URL
       let font = try loadFont(fromUrl: fontUrl, alias: fontFamilyAlias)
 
       if let postScriptName = font.postScriptName as? String {
         FontFamilyAliasManager.setAlias(fontFamilyAlias, forFont: postScriptName)
+        registeredFonts = Array(Set(registeredFonts).union([postScriptName, fontFamilyAlias]))
+      } else {
+        throw FontNoPostScriptException(fontFamilyAlias)
       }
-
-      registeredFonts = Array(Set(registeredFonts).union([fontFamilyAlias]))
     }
   }
 }